# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for Odinbit game",
        "title": "Odncore API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/license/mit"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/player/auth": {
            "post": {
                "description": "Register new player or login existing player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "player"
                ],
                "summary": "Authenticate player",
                "parameters": [
                    {
                        "description": "Player authentication data",
                        "name": "playerAuth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.PlayerAuth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "FAIL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/player/getpdata": {
            "get": {
                "description": "Get player's saved data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "Get player data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player nickname",
                        "name": "Session-Nickname",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/player.Player"
                        }
                    },
                    "404": {
                        "description": "Player not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/player/loadpdata": {
            "post": {
                "description": "Save player's current state",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "Save player data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player nickname",
                        "name": "Session-Nickname",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Player data to save",
                        "name": "playerData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/player.Player"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/server/info": {
            "get": {
                "description": "Get information about the server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "Get server info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ServerInfo"
                        }
                    }
                }
            }
        },
        "/server/status": {
            "get": {
                "description": "Get current server status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "Get server status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ServerStatus"
                        }
                    }
                }
            }
        },
        "/world/getworld": {
            "get": {
                "description": "Get current world state as binary data",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "world"
                ],
                "summary": "Get world data",
                "responses": {
                    "200": {
                        "description": "World binary data",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/world/loadid": {
            "post": {
                "description": "Load world identification data",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "world"
                ],
                "summary": "Load world ID",
                "parameters": [
                    {
                        "description": "World ID data",
                        "name": "idData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/world/loadworld": {
            "post": {
                "description": "Load world state data from binary file",
                "consumes": [
                    "application/octet-stream"
                ],
                "tags": [
                    "world"
                ],
                "summary": "Load world data",
                "parameters": [
                    {
                        "format": "binary",
                        "description": "World binary data",
                        "name": "worldData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.PlayerAuth": {
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "player.Player": {
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string"
                },
                "targetX": {
                    "type": "number"
                },
                "targetY": {
                    "type": "number"
                },
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
        "server.ServerInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "max_players": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "players_connected": {
                    "type": "integer"
                }
            }
        },
        "server.ServerStatus": {
            "type": "object",
            "properties": {
                "id_waiting": {
                    "type": "boolean"
                },
                "world_waiting": {
                    "type": "boolean"
                }
            }
        }
    },
    "tags": [
        {
            "description": "Server management endpoints",
            "name": "server"
        },
        {
            "description": "World management endpoints",
            "name": "world"
        },
        {
            "description": "Player management endpoints",
            "name": "player"
        }
    ]
}